lab6_inpaint/requirements.txt
matplotlib
torch
scikit_image
numpy
torchvision
pillow
pyamg
opencv-python
